Removed old water system and first water sim prototype
