--- Assets/_Project/Scripts/GridManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ed6efc9ce2d09cc43998c399fc02ab8b
