--- Assets/_Project/Scripts/InspectorController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e4f5f580db6643c438c187f11ca5472a
