--- Assets/_Project/Scripts/Old Water/Tile.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7306aab7576a770499820f8da266e96e
