--- Assets/_Project/Scripts/Old Water/WaterBlock.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d0abad6a66df48d4f8e78a5cc2b7fd3c
