--- Assets/_Project/Scripts/Old Water/WaterObststacle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2214e1ef88d414c4db8dfc77c6f852c2
