--- Assets/_Project/Scripts/Old Water/WaterTile.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 08f12a5e20f901f4189d9ca73e1e5776
