--- Assets/_Project/Scripts/Water System/Scriptable Object Scripts/FloodSimDataEditor.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 35b841f8339ed4c4685e6f699755d77f
